requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
behave==1.2.6
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
cssselect==1.2.0
disutils==1.4.32.post2
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.16.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
mycdp==1.1.0
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
pdbp==1.6.1
platformdirs==4.3.6
playwright==1.49.0
pluggy==1.5.0
pprintpp==0.4.0
propcache==0.2.1
py==1.11.0
pyee==12.0.0
Pygments==2.18.0
pynose==1.5.3
pyotp==2.9.0
PySocks==1.7.1
pytest==8.3.3
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
sbvirtualdisplay==1.3.0
selenium==4.27.1
seleniumbase==4.33.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tabcompleter==1.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
yarl==1.18.3
